Make footer a sticky footer
Somehow, this means "the footer should go at the bottom of the page even if
there isn't enough content above it for it to go that low" and not "create that
pervasive and annoying bottom popup that tells me about the existence of
cookies."

Don't make the always-visible footer. Do make the footer belong at the bottom
even if there isn't enough content to shove it down there.