Partially implement Mkdocs
This is a broken commit. I tried to update to Nanoc, but it turns out I'm so
unmotivated to touch nanoc that I stopped working on this for 4 months, so I'm
forcing the switch to MkDocs, which is Python and thus much easier to
understand. I'm not happy about this commit and it looks really janky, but
something something broken eggs omelet.